PBI: 9992032: [RVX] Enhance sample.html CR: See pull request.
I was having trouble getting started with RVX Widget because the
sample.html doesn't do anything. It still doesn't, after this change,
but at least it tells the user what three items he or she needs to start
streaming (with placeholders for two).

Changes
1) Added text boxes for user to enter client API endpoint, video name,
and JWT auth token for Client API. I pre-populated the first two to
play the rodrigb-toronto2 video from rodrigbanalyzer2. The third is a
secret and therefore must not be checked in (and will in any case
expire). If I had access to a public video which required no
authentication, I would have preferred to use that, but I don't know of
any.
2) Added a play button which will configure and load the RVX Widget with
the provided information. We also turn on debug mode so we can see RVX
widget traces in console.
3) Limited RVX widget to 720px width so I don't have to scroll down to
access the controls.

Testing Procedure
1) Start http-server and browse to sample.html. Obtain a client API auth
token for rodrigbanalyzer2, paste into JWT Auth Token for Client API
and press "Start Playback". Confirm that it plays rodrigb-toronto2.
2) Play a different video from a different account and confirm that
works, too.