PBI: 9993042: [RVX] Log additional media events CR: See pull request.
The current RVX Widget subscribes to some media events, but I needed to
subscribe to more in order to understand what was happening when
playback was failing to progress. Leaving this in place so that our logs
are more meaningful.

Changes
1) The checked-in shaka.ts is out of date. Added some new Shaka events
which I was interested in, such as onstatechange/StateChangeEvent,
but did NOT attempt to exhaustively add all missing Shaka events.
2) Added listeners to a set of interesting events (wrt media playback).
All these listeners do is to log the event.
3) While doing this, found and fixed two typos in shaka.ts:
loaded/LoadedEvent was incorrectly tied to loading/LoadingEvent,
gapSize was misspelled as gaSize.

Testing Procedure
1) Play a video from sample.html. Verify that you see events being
logged in console, such as loaded/loading status, buffering true ->
false, etc.