PBI: 9994070: [RVX] Expose shaka as global CR: See pull request.
We need to be able to access a global shaka object in order to set the
logging level, so that we can see what's going on in Shaka.

Changes
1) Currently, there are three places in the code which
require('shaka-player.ui.debug.js') (yes - the debug version), and
then keep this for local use.
2) Earlier versions of this fix just modified these require's in place
to also export them. This earlier fix worked fine, but I decided to
consolidate and require at a central position. The benefit is, if we
ever decide to switch to non-debug Shaka UI, we would only have to
change it in one place. So instead of three require's, I now do a single
require in index.ts.
3) Where to put the single require matters. Putting the require inside
player.class.ts, for instance, but also exporting from index.ts,
compiles fine but fails at runtime because shaka is not initialized when
index.ts tries to export it. So I am hoping that index.ts gets executed
first.
4) Consolidating three require's to one require does not appear to have
any effect on the .min.js output size.

Known Issues
1) PBI #9994215: [RVX] Consider shifting from shaka-player.ui.debug.js
to shaka-player.ui.js. The possible issues of doing so (and the
benefits) are described in this bug.

Testing Procedure
1) Open sample.html. In console, run something like
shaka.log.setLevel(shaka.log.Level.V2) before and/or after you start
video playback and confirm that it has an effect on the amount of log
messages being sent to console. Prior to this change, such a statement
would fail because “shaka” == undefined.