Bug 9963755: [AVA] Shaka hangs on short discontinuous segment playback CR: See pull request.
When playing on-demand or live content, Shaka can hang indefinitely
sometimes. What we have found is that we can often resume playback by
seeking Shaka.

Changes
1) Check every 3 seconds whether or not Shaka is stalled. We use a 1 ms
threshold on HTMLMediaElement.currentTime, and we also check if
HTMLVideoElement is paused.
2) If Shaka is stalled but either Shaka or HTMLVideoElement thinks that
there is sufficient media to begin playback, then we seek Shaka to
current position + 1 second to attempt to break the impasse. Oftentimes
+ 0 or + 0.1 is enough, but where this can fail is when there are
discontinuities in the media, so +1 allows us to get past these
discontinuities faster and only loses a little user media.
3) In the live case, Shaka really is buffering, so instead we verify
that we were playing in live position before attempting a seek to
live, but only after 9 seconds of failure to progress (because we want
to give time for buffering to complete). It is important not to seek to
live if user was actually playing in DVR as opposed to live position.
4) Finally, if we encounter 10 consecutive stall detections, we attempt
a seek to same (+0 seconds) as this can sometimes break an impasse
and is low-risk.

Testing Procedure
1) Hang at start: play rodrigb-vehicle-counting-cvr-007 from
rodrigbanalyzer2 in RVX widget. Confirm this no longer stalls at
start of playback.
2) Hang mid-stream and start: With the new default Shaka bufferingGoal
of 30, Test001-AxisCamera from rodrigbanalyzer2 no longer stalls at
start or midstream. To induce stall, modify player.class.ts in RVX
codebase so that bufferingGoal is 10 instead of 30. You should stall at
start, and burn-in  times of 21:19:50 and 21:34:04 mid-stream.
3) Hang mid-stream: Play rodrigb-toronto from rodrigbanalyz, seek to
around mid-way, burn-in time should be 19:50 or so (I forget the
hour).
4) Live hang: play simple-audio-video from krishnduuswc, or else your
own live camera feed. While live streaming (~45 seconds behind live),
log into camera webpage and restart camera and/or change RTSP password
to lengthen the outage (did not test the latter). The playback should
stall (no way to avoid, no media being produced), but should eventually
recover. We generally see it recover and stall around three times before
permanently recovering. Prior to this fix, it would stall indefinitely.