BUG: 10009964: RVX: Player seek in VOD Timeline for Live assets can still cause mismatched seek positions CR: (pull request).
An earlier fix to getVideoOffset() was contributed to fix a hang when
transitioning from live view to on-demand. When the fix was published,
we tested it against all our available live feeds, and we found that the
longer a camera feed has been up, the more broken seeking becomes in the
on-demand view. It starts off seeking somewhere other than where you
seeked (typically, to the left/earlier), and eventually degenerates to
the point where clicking anywhere on the timeline view only seeks to the
extreme left.

Changes
1) The new fix fixes the old hang scenario AND the new seek error. It is
conceptually simple: it is never OK to seek video.currentTime outside
of [shaka]player.seekRange() start/end. To achieve this, we map the
start time of the first available-media segment to the seekRange() start
time. Fixed getVideoOffset() to carry out this mapping. Because the
available-media segments are not currently being saved where they can be
accessed, I added a new class member to store the start time in seconds
of the first segment.
2) Bumped the package.json version from 0.4.8 to 0.4.9. This causes an
update to package-lock.json when npm install is run, so I am checking
in that file along with package.json (as should any other developer who
updates package.json, but it seems this has not been happening).
3) While updating package.json, I realized that the shaka-player
dependency specifies ^3.1.0, which allows for auto-update, and
Prakash reminded me that 3.1.1 is due out soon. I don’t want the risk of
regression so I removed the ^ to lock to 3.1.0 specifically.
4) Enhanced the existing video.currentTime seek log messages to also
print the valid seek range, which is needed to interpret whether the
seek position was as expected.

Details
1) This earlier fix made the assumption that the DASH manifest from
origin would follow DASH spec (https://dashif-documents.
azurewebsites.net/Guidelines-TimingModel/master/Guidelines-TimingModel.
html#period-timing). It does not, because it can return a
presentationTimeOffset from days ago, yet still have first period start
time of 0, resulting in a seek range starting at a value other than 0.
This is not DASH compliant and some players could fail.

Known Issues
1) Bug # 10014971: AVA: Origin live manifest is not DASH compliant.
2) The fact that RVX Widget uses availability ranges from the Rest API,
instead of data from Shaka, may lead to issues in future. For
example, early versions of this fix had an alarming regression where the
rodrigb-toronto2 clip from rodrigbanalyzer2, would loop infinitely in
the first 3 seconds. We assumed that the first segment returned by
createTimelineSegments() represented the first availability segment, but
createTimelineSegments() discards segments < 5 seconds. Thus we would
seek backwards instead of forwards to the next segment. This never would
have happened if using data from Shaka.

Testing Procedure
1) Test against a matrix of camera feeds WITH time burn-in which: a)
started at varying times (eg. 8 days ago, 1 day ago, just now), b)
have or do not have media at 00:00 GMT today, c) have and do not have
discontinuities today and in previous days, etc. Play against all feeds,
toggle between live/on-demand today/on-demand previous days. In all
cases, seek around and verify that the seek went where you expected to
go by looking at the time burn-in (the timeline markings are excellent
for this as they are on the hour). DO NOT use the playback cursor as
"proof" because we have seen cases where the playback cursor can be
painted in the wrong place (typically, extreme right), but playback is
actually still on the timeline.