PBI: 9992032: [RVX] Enhance sample.html, part 2. CR: See pull request.
Prakash recommended some additional enhancements to sample.html.

Changes
1) Prakash suggested that sample.html should use global.js, not index.js
(formerly ava-player.js), because the latter was intended for node
usage, and the former was meant to be used by browsers from CDN and
defines a top-level ava variable, so that you can immediately start
referring to ava.shakaPlayer, etc. Unfortunately, this ava object seems
to be different than the one which is actually performing playback.
2) I was noticing an error in console whenever loading sample.html. It
was complaining, "Uncaught ReferenceError: onAVALoad is not defined
at HTMLScriptElement.onload". Fixed by rearranging the <script> blocks
so that onAVALoad is defined when referenced in HTML.

Testing Procedure
1) Browse to sample.html. In developer console, verify that the
top-level ava variable is no longer undefined. Unfortunately, it
doesn’t seem to be the playback ava player.