FEAT 10619548: Handle MSE errors by resetting source, part 1. CR: See pull request.
The long-term home for this retry logic will be in the AvaPlayer class
of the ShakaRTSP repo, but until AvaWidgets moves to AvaPlayer, the code
which will go there needs to temporarily go in here first. We therefore
expect a part 2 commit which will undo this one.

Changes
1) When we receive a Shaka error callback, if we are playing live and
the error code is 3016 (VIDEO_ERROR, which we are commonly seeing as
VDA Error/PIPELINE_ERROR_DECODE, we will re-load the Shaka player up to
10 times within a three-minute window. This is generally enough to keep
going for quite some time, even in the face of continual decode
errors (I was able to go 25 minutes, for instance). The danger is if
buffering times are long (> 18 seconds), then 10 times in 3 minutes
could end up meaning infinite retry. We do not retry when playing
on-demand and for any other error codes than 3016.
2) As part of the error handling, we had to suppress further propagation
of the error event by calling stopImmediatePropagation (otherwise,
for example, the upper layers would stop playback and display an error
message). Extended the shaka.ts declarations to allow for this.
Long-term we should delete this file and use the declarations which ship
with Shaka.

Testing Procedure
1) Set up AVA endpoint against a b-frame camera stream (eg. Bosch). If
possible, modify the ShakaRTSP/media-stream-library to under-profile
the AVC1 codec to increase the chances of decode error. Run sample.html
against the AVA endpoint and confirm from console logs that we will
retry up to 10 times in a 3-minute window, and if we exceed this, we
stop trying and allow the error event to propagate, resulting in
playback stoppage and error message display.