PBI: 11112223: Temp error handler/reconnect for Widgets, part 1. CR: See pull request.
Originally, I had intended to have Widgets call ShakaRTSP's error
handler/reconnect logic, which was added in PBI # 11064018. It turns
out, this would have been the first time that Widgets had made calls
into ShakaRTSP, and a number of blocking issues arose. Since these
issues will take time to resolve, the quick fix is to simply paste the
code into player.class.ts, and then remove it later when we have
resolved the blocking issues.

Changes
1) Paste the "part 2" version of PBI # 11064018's error handler and
reconnect logic into player.class.ts. During the course of
integration, it was discovered that the return value of
shakaErrorHandler.onShakaError was not useful (it would return true to
indicate that the error was recognized). What we really needed was to
know that it was handled, which is now true with the part 2 change. As
part of this, had to update definitions in shaka.ts. Additionally, had
to modify the code to meet the different lint requirements in this repo.
2) Note that his commit relies on ShakaRTSP version 1.0.4 in order to
receive the newer shaka errors being emitted from
rtspManifestParser.ts.

Testing Procedure
1) Play against b-frame AVA endpoint on Chrome with hardware
acceleration. Filter console log for 3016 (decode error). Confirm
that after 3016, we reconnect and keep playing. This was still true
prior to this change, but re-testing is to confirm that we have not
regressed.
2) Play until server disconnects (or induce temporary network failure).
Restore network. Confirm that we reconnect and keep playing. Prior to
this change, we did not automatically reconnect. Note that this differs
from failure during initial websocket connection: this change does not
modify that. If we encounter failure during initial websocket connection
(bad credentials, bad hostname, bad IP, etc), regardless of this commit,
we will retry 5 times before giving up.